102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	int a = 1, b = 2, c,  i;
+	unsigned long a = 1, b = 2, c,  i;
 
 	for (i = 1; i <= 50; i++)
 	{
